app/requirements.txt
"fastapi[all]"
pydub
google-cloud-speech